And-13 логика получения счёта за потреблённые услуги

+++ domain/src/main/kotlin/ru/surfstudio/standard/domain/entity/Payment.kt
@@ -2,5 +2,5 @@ package ru.surfstudio.standard.domain.entity
 
 data class Payment(
     val type: String,
-    val value: Float
+    val value: Int
 )
